fix(checkcode): 修复 PHP 8.3 浮点坐标转换异常
imagettftext() 的 x/y 坐标参数在 PHP 8.3 中要求 int 类型，对除法和乘法运算结果加 (int) 强制转换。

+++ tp6/app/common/libs/checkcode/Checkcode.php
@@ -106,7 +106,7 @@ class Checkcode
             if (!$isFontcolor) {
                 $this->fontcolor = imagecolorallocate($this->img, mt_rand(0, 156), mt_rand(0, 156), mt_rand(0, 156));
             }
-            imagettftext($this->img, $this->fontsize, mt_rand(-30, 30), $_x * $i + mt_rand(1, 5), $this->height / 1.4, $this->fontcolor, $this->font, $this->code[$i]);
+            imagettftext($this->img, $this->fontsize, mt_rand(-30, 30), (int) ($_x * $i + mt_rand(1, 5)), (int) ($this->height / 1.4), $this->fontcolor, $this->font, $this->code[$i]);
         }
     }
